+++ .eslintrc.js
@@ -42,6 +42,8 @@ module.exports = {
           {argsIgnorePattern: '^_'},
         ],
         '@typescript-eslint/prefer-ts-expect-error': 'error',
+        // TS verifies this
+        'consistent-return': 'off',
         // Since we do `export =`. Remove for Jest 27
         'import/default': 'off',
         'no-dupe-class-members': 'off',