+++ packages/jest-haste-map/src/index.ts
@@ -325,6 +325,7 @@ export default class HasteMap extends EventEmitter {
       (options.ignorePattern || '').toString(),
       hasteImplHash,
       dependencyExtractorHash,
+      this._options.computeDependencies.toString(),
     );
     this._buildPromise = null;
     this._watchers = [];