+++ packages/jest-types/src/Config.ts
@@ -196,6 +196,9 @@ export type InitialOptions = Partial<{
   onlyFailures: boolean;
   outputFile: Path;
   passWithNoTests: boolean;
+  /**
+   * @deprecated Use `transformIgnorePatterns` options instead.
+   */
   preprocessorIgnorePatterns: Array<Glob>;
   preset: string | null | undefined;
   prettierPath: string | null | undefined;
@@ -209,8 +212,14 @@ export type InitialOptions = Partial<{
   roots: Array<Path>;
   runner: string;
   runTestsByPath: boolean;
+  /**
+   * @deprecated Use `transform` options instead.
+   */
   scriptPreprocessor: string;
   setupFiles: Array<Path>;
+  /**
+   * @deprecated Use `setupFilesAfterEnv` options instead.
+   */
   setupTestFrameworkScriptFile: Path;
   setupFilesAfterEnv: Array<Path>;
   silent: boolean;
@@ -227,6 +236,9 @@ export type InitialOptions = Partial<{
   testLocationInResults: boolean;
   testMatch: Array<Glob>;
   testNamePattern: string;
+  /**
+   * @deprecated Use `roots` options instead.
+   */
   testPathDirs: Array<Path>;
   testPathIgnorePatterns: Array<string>;
   testRegex: string | Array<string>;
@@ -462,6 +474,7 @@ export type Argv = Arguments<
     silent: boolean;
     snapshotSerializers: Array<string>;
     testEnvironment: string;
+    testEnvironmentOptions: string;
     testFailureExitCode: string | null | undefined;
     testMatch: Array<string>;
     testNamePattern: string;